// src/types/generated/MatchDiscoveryPlan.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { MatchDiscoveryDecision } from "./MatchDiscoveryDecision";
import type { PlayerCrawlDecision } from "./PlayerCrawlDecision";

export type MatchDiscoveryPlan = { to_crawl: Array<string>, new_match_ids: Array<string>, player_decisions: Array<PlayerCrawlDecision>, match_decisions: Array<MatchDiscoveryDecision>, selected_crawl_count: number, new_match_count: number, skipped_count: number, };